migration to cairo 1.18
Still a couple of changes required compared to official 1.18.4. Patches included.

+++ util/cairo-script/cairo-script-scanner.c
@@ -46,7 +46,7 @@
 #include <zlib.h>
 
 #if HAVE_LZO
-#include <lzo/lzo2a.h>
+#include <lzo2a.h>
 #endif
 
 #define DEBUG_SCAN 0
@@ -1054,6 +1054,7 @@ scan_token:
 	switch (c) {
 	case 0xa:
 	    scan->line_number++;
+	    /* fall through */
 	case 0x0:
 	case 0x9:
 	case 0xc:
@@ -1558,7 +1559,7 @@ _translate_real (csi_t *ctx,
 {
     uint8_t hdr;
 
-    if (real >= INT32_MIN && real <= INT32_MAX && (int) real == real)
+    if ((double)real >= INT32_MIN && (double)real <= INT32_MAX && (int) real == real)
 	return _translate_integer (ctx, real, closure);
 
 #if WORDS_BIGENDIAN
